Implemented Square and Triangle Grids.
The Vector & Grid traits exist to allow a variety of Grids. However with only Hexagonal grids implemented it was unclear if the implementation could support even all regular grids. Therefore adding implementations of Triangle and Square grids gives confidence in the implementation. The Hex tests have been duplicated to work for both grids and succeed.