Swapped Hexagon grid to a radius-based hexagonal shape as with sirpent-go.
This is for 46bit's convenience and might be a precursor to world shapes - types defining a particular shape of world as used in 0.1 of the JSON protocol. Sending AIs a list of extant cells is also an option but one I feel less comfortable with for now. Lots of real experimentation is needed.