Major rework to accomodate new Command and Message classes, including (de)serialising Directions and decoding JSON directly from the Stream with no need for specified message lengths.
The Command enum is based upon https://github.com/aatxe/irc/blob/master/src/client/data/command.rs. Requiring a Vector type to be specified when decoding any Command is unfortunate. I wonder if there's a way to remove that, without weakening the Direction type system.

This uses the new Server setup for #3, but primarily is for #8.