Experimental asyncronous fork using Tokio.
Sirpent-rust has been a pain to work with. The game engine and network communication has been quite the struggle to code in a Rustic manner. This is partly because of a lack of examples of how to build network programs that aren't simple Request-Response servers.

Programming in Rust as if it were C/C++ with big formalised state machines seems to discard a lot of the benefits. So I've been trying an event-driven approach.

I experimented with mio but found it too low-level to readily build upon. Tokio wraps around mio using Futures and those seem to have a lot of potential to simplify sirpent-rust.