Made CauseOfDeath less verbose; wrapped Grid with a tiling label for Msgs.
CauseOfDeath's verbosity wasn't particularly necessary. Clients smart enough to care about those details can easily calculate the information for themselves. The protocol is easier to implement in typed languages with a simple string for each `Msg::Outcome.conclusion.casualties`: `no_move_made`, `collided_with_snake` or `collided_with_bounds`.

Unfortunately munged in is another change. I've added a `GridEnum` that labels grids with their tiling when sent as JSON. This doesn't work around type system limitations (the compile-flag of which grid to use remains) but outside of Rust it's going to be easy to be generic over known grids with this. In Rust it makes it easy to disconnect if needed.